backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "graphify-rag-backend"
version = "0.1.0"
description = "Standalone RAG backend for document knowledge bases."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.4.0",
    "openai>=1.60.0",
    "qdrant-client>=1.11.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
